+++ CHANGELOG.md
@@ -10,6 +10,10 @@
 - You can now use _ or - in volt tag names and attributes.  (We're moving to using dash ( - ) as the standard in html)
 - You can now trigger events on controllers rendered as tags.  The events will bubble up through the DOM and can be caught by any e- bindings.  See the docs for more information.
 
+### Changed
+- fix issue with ```raw``` and promises (#275)
+- fix issue with .length on store (#269)
+
 ## 0.9.4
 ### Lingo Change
 the base collections will now be called "Repositories" or "Repo's" for short.  This will only matter directly for internal volt code, but for the data provider api, this will help.